rename ClickOption to ClickMethod; rename ClickParams to ClickOptions
